Write documentation for new `canvas` module
