+++ wgpu/src/widget/canvas/fill.rs
@@ -1,7 +1,9 @@
 use iced_native::Color;
 
+/// The style used to fill geometry.
 #[derive(Debug, Clone, Copy)]
 pub enum Fill {
+    /// Fill with a color.
     Color(Color),
 }
