+++ wgpu/src/widget/canvas/stroke.rs
@@ -1,10 +1,16 @@
 use iced_native::Color;
 
+/// The style of a stroke.
 #[derive(Debug, Clone, Copy)]
 pub struct Stroke {
+    /// The color of the stroke.
     pub color: Color,
+    /// The distance between the two edges of the stroke.
     pub width: f32,
+    /// The shape to be used at the end of open subpaths when they are stroked.
     pub line_cap: LineCap,
+    /// The shape to be used at the corners of paths or basic shapes when they
+    /// are stroked.
     pub line_join: LineJoin,
 }
 
@@ -19,10 +25,16 @@ impl Default for Stroke {
     }
 }
 
+/// The shape used at the end of open subpaths when they are stroked.
 #[derive(Debug, Clone, Copy)]
 pub enum LineCap {
+    /// The stroke for each sub-path does not extend beyond its two endpoints.
     Butt,
+    /// At the end of each sub-path, the shape representing the stroke will be
+    /// extended by a square.
     Square,
+    /// At the end of each sub-path, the shape representing the stroke will be
+    /// extended by a semicircle.
     Round,
 }
 
@@ -42,10 +54,15 @@ impl From<LineCap> for lyon::tessellation::LineCap {
     }
 }
 
+/// The shape used at the corners of paths or basic shapes when they are
+/// stroked.
 #[derive(Debug, Clone, Copy)]
 pub enum LineJoin {
+    /// A sharp corner.
     Miter,
+    /// A round corner.
     Round,
+    /// A bevelled corner.
     Bevel,
 }
